Update UI tests
