--- scorecard/tests/test_scorecard.py
@@ -1,16 +1,0 @@
-from django.contrib.sites.models import Site
-from django.core.management import call_command
-
-from municipal_finance.tests.helpers import BaseSeleniumTestCase
-
-
-class ScorecardTest(BaseSeleniumTestCase):
-    fixtures = ["seeddata", "demo-data", "compiled_profile"]
-
-    def setUp(self):
-        super(ScorecardTest, self).setUp()
-        Site.objects.filter(id=2).update(domain='municipalmoney.org.za', name='Scorecard')
-
-    def test_scorecard_formula(self):
-        selenium = self.selenium
-        selenium.get("%s%s" % (self.live_server_url, "/profiles/municipality-BUF-buffalo-city/"))
